+++ backend/src/main/java/spring/crut/corpus/repositories/info/InfoRepository.java
@@ -7,6 +7,6 @@ import spring.crut.corpus.models.info.Info;
 import java.util.Optional;
 
 @NoRepositoryBean
-public interface InfoRepository <T extends Info> extends JpaRepository<T, Integer> {
+public interface InfoRepository <T extends Info> extends JpaRepository<T, Long> {
     Optional<T> findByName(String name);
 }
